블럭과 연결된 다른 블럭들을 기록하도록 적용

+++ iceball/src/lib.rs
@@ -48,7 +48,7 @@ pub enum Memory {
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub enum RelativeAddressingArgument {
     Register(Register),
-    Constant(u64),
+    Constant(i128),
     Operator(AddressingOperator),
 }
